Dedicated book container and conditinal innerHTML clearing
Added dedicated .book-container to the HTML to prevent the element not existing error
Check if bookContainer exists, if true clear the bookContainer innerHTML to prevent duplicate entries. If false log an error to the console and return to break out of the function.
Updated myLibrary.forEach to appendChild to the bookContainer element instead of document.body